+++ apps/ymir-sdl3/src/app/ui/widgets/input_widgets.hpp
@@ -26,6 +26,7 @@ private:
     UnboundActionsWidget &m_unboundActionsWidget;
 
     void CaptureButton(input::InputBind &bind, size_t elementIndex, void *context);
+    void CaptureTrigger(input::InputBind &bind, size_t elementIndex, void *context);
     void CaptureAxis1D(input::InputBind &bind, size_t elementIndex, void *context);
     void CaptureAxis2D(input::InputBind &bind, size_t elementIndex, void *context);