+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 - CD Block: Fix handling of "no change" PlayDisc parameters. Fixes X-Men: Children of the Atom CDDA tracks not resuming after pausing. (#274)
 - Debugger: Indirect SCU DMA transfers were being traced with the updated indirect table address.
+- Input: Fix inability to bind keyboard combos.
 - Input: Modifier keys can now be used correctly as controller input binds and will no longer interfere with other controller inputs. (#282)
 - Media: Skip blank lines in CUE files.
 - SCSP: Use EG level instead of total level in MSLC reads. Fixes missing/truncated SFX on various games, including Sonic R, Akumajou Dracula X and Daytona USA CCE.